moved all methods to bot/mod.rs
