+++ .idea/teleborg-1.iml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$" />
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Rust &lt;teleborg-1&gt;" level="project" />
+    <orderEntry type="library" name="Cargo &lt;teleborg-1&gt;" level="project" />
+  </component>
+</module>
